+++ CHANGES.rst
@@ -4,8 +4,9 @@ Products.MeetingPROVHainaut Changelog
 4.2.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Fixed instance not starting because `Products.MeetingCommunes` `examples_fr`
+  and `zones` profiles were changed.
+  [gbastien]
 
 4.2.1 (2024-03-14)
 ------------------